+++ src/graphics/effects/Fissure.h
@@ -116,7 +116,7 @@ private:
 	bool bIntro;
 	
 	float sizeF;
-	float m_visibleNotches;
+	int m_visibleNotches;
 	float tfRaysa[40];
 	float tfRaysb[40];
 	Vec3f va[40];
